stage1_classifier/requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
seaborn>=0.13.0
matplotlib>=3.8.0
feedparser>=6.0.0
